Mudança de organização das views
