+++ server/api/test_views.py
@@ -3,7 +3,7 @@ from django.core import mail
 from rest_framework.test import APIClient
 from rest_framework.test import APIRequestFactory
 from rest_framework import status
-from .models import * 
+from .models import *
 from rest_framework.test import force_authenticate
 from model_bakery import baker
 # Create your tests here.
@@ -14,19 +14,14 @@ class EmailViewsTest(TestCase):
     def setUpClass(cls):
         super().setUpClass()
         poll = baker.make('api.Poll')
-        group = baker.make('api.Group')
         poll.id = 1
-        poll.name = 'Eleicao Teste'
+        poll.title = 'Eleicao Teste'
         cls.poll_id = 1
         poll.save()
-        group.poll_set.add(poll)
         for n in range(1, 6):
             user = baker.make('api.UserAccount')
             user.ref = 'user{}@abc.de'.format(n)
             user.save()
-
-            #Add the user in the poll:
-            poll.group.users.add(user)
 
             #Now, create the token:
             Token.objects.create_token(poll_id=poll.id, email=user.ref)
@@ -35,25 +30,19 @@ class EmailViewsTest(TestCase):
             user = baker.make('api.UserAccount')
             user.ref = 'user_without_token{}@abc.de'.format(n)
             user.save()
-            #Add the user in the poll:
-            poll.group.users.add(user)
         #Create users that has already voted:
         for n in range(1, 6):
             user = baker.make('api.UserAccount')
             user.ref = 'user_that_has_voted{}@abc.de'.format(n)
             user.save()
-            #Add the user in the poll:
-            poll.group.users.add(user)
             #Add the user in the list emails_voted
             poll.emails_voted.add(user)
         #Create users that are not active.
         for n in range(1, 6):
             user = baker.make('api.UserAccount')
             user.ref = 'not_active_user{}@abc.de'.format(n)
-            user.is_active = False 
+            user.is_active = False
             user.save()
-            #Add the user in the poll:
-            poll.group.users.add(user)
             #Now, create the token:
             Token.objects.create_token(poll_id=poll.id, email=user.ref)
 
@@ -65,11 +54,11 @@ class EmailViewsTest(TestCase):
         poll = Poll.objects.get(id=self.poll_id)
         user = poll.admin
         client.force_authenticate(user=user) #Do not check authentication
-        response = client.post('/api/poll/send-poll-emails', {'poll_id':poll.id}, format='json')
+        response = client.post('/api/emails/send', {'poll_id':poll.id}, format='json')
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(mail.outbox), 5)
         for i in range(5):
-            self.assertEqual(mail.outbox[i].subject, 'Link para a eleicao {}'.format(poll.name))
+            self.assertEqual(mail.outbox[i].subject, 'Link para a eleicao {}'.format(poll.title))
 
     #'send_emails' test
     def test_send_list_emails__http_response(self):
@@ -91,22 +80,13 @@ class EmailViewsTest(TestCase):
         Token.objects.create_token(poll_id=poll.id, email=new_user2.ref)
         Token.objects.create_token(poll_id=poll.id, email=new_user3.ref)
 
-        response = client.post('/api/poll/send-list-emails', {'poll_id':poll.id, 'users_id_list':[new_user1.id, new_user2.id]}, format='json')
+        response = client.post('/api/emails/send-list', {'poll_id':poll.id, 'users_id_list':[new_user1.id, new_user2.id]}, format='json')
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(mail.outbox), 2)
-        self.assertEqual(mail.outbox[0].subject, 'Link para a eleicao {}'.format(poll.name))
-        self.assertEqual(mail.outbox[1].subject, 'Link para a eleicao {}'.format(poll.name))
+        self.assertEqual(mail.outbox[0].subject, 'Link para a eleicao {}'.format(poll.title))
+        self.assertEqual(mail.outbox[1].subject, 'Link para a eleicao {}'.format(poll.title))
 
-        response = client.post('/api/poll/send-list-emails', {'poll_id':poll.id, 'users_id_list':[new_user3.id]}, format='json')
+        response = client.post('/api/emails/send-list', {'poll_id':poll.id, 'users_id_list':[new_user3.id]}, format='json')
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(mail.outbox), 3)
-        self.assertEqual(mail.outbox[2].subject, 'Link para a eleicao {}'.format(poll.name))
-
-
-
-
-
-
-
-
-
+        self.assertEqual(mail.outbox[2].subject, 'Link para a eleicao {}'.format(poll.title))
