+++ server/api/urls.py
@@ -3,17 +3,23 @@ from django.urls import path
 from rest_framework_simplejwt import views as jwt_views
 
 urlpatterns = [
-    path('hello', HelloView.as_view()),
-    path('poll/send-list-emails', send_list_emails),
-    path('poll/send-poll-emails', send_poll_emails),
-    path('poll/create', create_poll),
-    path('poll/return-result', return_result),
-    path('poll/delete/<int:pk>/', delete_poll),
-    path('group/create', create_group),
+    # Padrão -> /api/<nome do recurso>/<operação>
+
+    path('polls/get/<int:pk>/', get_poll),
+    path('polls/create', create_poll),
+    path('polls/update/<int:pk>', update_poll),
+    path('polls/delete/<int:pk>/', delete_poll),
+    path('polls/mine', get_user_polls),
+    path('polls/get/<int:pk>/result', get_poll_result),
+
+    path('emails/send', send_poll_emails),
+    path('emails/send-list', send_list_emails),
+
+    path('groups/create', create_group),
     path('groups/mine', user_groups),
+
     path('tokens/create', register_emails),
-    path('tokens/create_from_group', register_emails_from_group),
-    path('vote/compute', compute_vote),
-    path('polls/mine', user_polls),
-    path('polls/options/<int:pk>/', poll_options)
+    path('tokens/create-from-group', register_emails_from_group),
+
+    path('votes/compute', compute_vote),
 ]
