+++ server/api/views/__init__.py
@@ -1,0 +1,5 @@
+from .emails import *
+from .polls import *
+from .groups import *
+from .tokens import *
+from .votes import *
