+++ src/styles/poll-view.js
@@ -1,0 +1,29 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles({
+    root: {
+        width: '40%',
+        justifyContent: 'center',
+        textAlign: 'center',
+        marginLeft: '30%',
+        marginTop: '5%'
+    },
+    bullet: {
+        display: 'inline-block',
+        margin: '0 2px',
+        transform: 'scale(0.8)',
+    },
+    title: {
+        fontSize: 14,
+    },
+    pos: {
+        marginBottom: 12,
+    },
+    paper: {
+        height: '100%',
+        verticalAlign: 'middle',
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center"
+    },
+});
